+++ src/gleam_stdlib.erl
@@ -10,7 +10,7 @@
          percent_encode/1, percent_decode/1, regex_check/2, regex_split/2,
          base_decode64/1, parse_query/1, bit_string_concat/1, size_of_tuple/1,
          decode_tuple/1, tuple_get/2, classify_dynamic/1, print/1, println/1,
-         inspect/1]).
+         inspect/1, pow/2]).
 
 %% Taken from OTP's uri_string module
 -define(DEC2HEX(X),
@@ -375,3 +375,8 @@ inspect(Any) when is_function(Any) ->
     ["//fn(", Args, ") { ... }"];
 inspect(Any) ->
     ["//erl(", io_lib:format("~p", [Any]), ")"].
+
+pow(X, Y) ->
+    try {ok, math:pow(X, Y)}
+    catch error:badarith -> {error, nil}
+    end.